octomap: do not hardcode absolute paths in octomap-config.cmake
Hardcoding abslotue install paths in octomap-config.cmake totally breaks
the cross-compilation: it makes impossible to be used in a sysroot because
the hardcoded paths are the ones of the target, whereas they should point
to the sysroot location.

This patch uses the CMake find_library and find_path to set the variables
OCTOMAP_LIBRARIES and OCTOMAP_INCLUDE.
It also get rid of the unused OCTOMAP_LIBRARY_DIR variable.