dynamicEDT3D: do not hardcode absolute paths in dynamicEDT3DConfig.cmake
Hardcoding abslotue install paths in octomap-config.cmake totally breaks
the cross-compilation: it makes impossible to be used in a sysroot because
the hardcoded paths are the ones of the target, whereas they should point
to the sysroot location.

This patch uses the CMake find_library and find_path to set the variables
DYNAMICEDT3D_LIBRARIES and DYNAMICEDT3D_INCLUDE_DIRS.
It also get rid of the unused DYNAMICEDT3D_LIBRARY_DIRS variable.